# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Workshops API",
        "description": "REST API для управления мастер-классами, площадками и регистрациями. Вариант 15.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "paths": {
...
        "/workshops": {
            "get": {
                "tags": [
                    "Мастер-классы"
                ],
                "summary": "Список мастер-классов с фильтрами",
                "operationId": "c40ca0df753176004d52fc9f4c1fad87",
                "parameters": [
                    {
                        "name": "organizer_id",
                        "in": "query",
                        "description": "Фильтр по организатору",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "С даты",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "has_available_spots",
                        "in": "query",
                        "description": "Только с местами",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка (-starts_at, price, title)",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список с пагинацией"
                    }
                }
            },
            "post": {
                "tags": [
                    "Мастер-классы"
                ],
                "summary": "Создать мастер-класс (организатор/админ)",
                "operationId": "fc091eada51a654b90f223e0b1cb8cb0",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "venue_id",
                                    "title",
                                    "starts_at",
                                    "ends_at",
                                    "max_participants"
                                ],
                                "properties": {
                                    "venue_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Мастер-класс по рисованию"
                                    },
                                    "starts_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-12-25 10:00:00"
                                    },
                                    "ends_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-12-25 13:00:00"
                                    },
                                    "max_participants": {
                                        "type": "integer",
                                        "example": 20
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 1500
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Создано"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/workshops/{id}": {
            "get": {
                "tags": [
                    "Мастер-классы"
                ],
                "summary": "Просмотр мастер-класса",
                "operationId": "b702494f8c6daccf975238aa1f10128f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные"
                    },
                    "404": {
                        "description": "Не найдено"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Мастер-классы"
                ],
                "summary": "Удалить мастер-класс (владелец/админ)",
                "operationId": "18cfdefb992b9166d26b1cdd92b58a51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Удалено"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Мастер-классы"
                ],
                "summary": "Обновить мастер-класс (владелец/админ)",
                "operationId": "97bd47b9350a0d6391b23eff5803cb7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Новое название"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Обновлено"
                    },
                    "403": {
                        "description": "Нет прав"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum",
                "description": "Вставьте токен в формате: Bearer {token}"
            }
        }
    },
...
}